==> notebooks/training/04_predict-stress.ipynb <==
# 04 - Battery Stress Prediction

This notebook demonstrates how to use the trained stress detection model to predict whether a battery is under stress based on new input data.


## Import Required Libraries

Import the necessary libraries for loading the model and making predictions.


import os
import warnings

os.environ['CUDA_VISIBLE_DEVICES'] = '-1'
os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'

warnings.filterwarnings('ignore', category=UserWarning, module='google.protobuf')
warnings.filterwarnings('ignore', category=DeprecationWarning)

import numpy as np
import pandas as pd
import tensorflow as tf
import joblib

## Load Model and Prepare Test Data

Load the trained stress detection model and scaler. Create a sample battery reading with stressed conditions (high current, high temperature) for testing.


model = tf.keras.models.load_model("./models/battery_stress_model.keras")
scaler = joblib.load("./models/stress_scaler.pkl")

new_data = pd.DataFrame([{
    "stateOfCharge": 0.15,
    "stateOfHealth": 97.00,
    "batteryCurrent": 480.00,
    "batteryVoltage": 355.00,
    "kmh": 150.00,
    "distance": 50.00,
    "batteryTemp": 55.00,
    "ambientTemp": 22.00,
    "currentLoad": 100,
}])

input_data = scaler.transform(new_data).astype(np.float32)

## Make Prediction

Use the loaded model to predict whether the battery is stressed. The output is a probability that is converted to a binary classification (STRESSED or NORMAL).


prediction = model.predict(input_data, verbose=0)
predicted_probability = prediction[0][0]
predicted_class = 1 if predicted_probability > 0.5 else 0

print("Stress Prediction:", "STRESSED" if predicted_class == 1 else "NORMAL")
