==> notebooks/training/06_predict-ttf.ipynb <==
# 06 - Time-to-Failure Prediction

This notebook demonstrates how to use the trained time-to-failure model to predict how many hours remain before a battery failure based on current battery conditions.


## Import Required Libraries

Import the necessary libraries for loading the model and making predictions.


import os
import warnings

os.environ['CUDA_VISIBLE_DEVICES'] = '-1'
os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'

warnings.filterwarnings('ignore', category=UserWarning, module='google.protobuf')
warnings.filterwarnings('ignore', category=DeprecationWarning)
warnings.filterwarnings('ignore', category=UserWarning, module='keras')
warnings.filterwarnings('ignore', category=UserWarning, module='sklearn')

import numpy as np
import tensorflow as tf
import joblib

## Load Model and Prepare Test Data

Load the trained TTF model and scaler. Create a sample battery reading with specific temperature, current, voltage, state of charge, and state of health values.


model = tf.keras.models.load_model("./models/battery_ttf_model.keras")
scaler = joblib.load("./models/ttf_scaler.pkl")

raw_data = np.array([[45.0, 500.0, 370.0, 0.4, 98.0]])
input_data = scaler.transform(raw_data).astype(np.float32)

## Make Prediction

Use the loaded model to predict the remaining time before battery failure in hours.


prediction = model.predict(input_data, verbose=0)
predicted_ttf = prediction[0][0]
print(f"Predicted Time Before Failure: {predicted_ttf:.2f} hours")
